Added trait to make derived const legible to the type system
