+++ Cargo.toml
@@ -7,13 +7,6 @@ license-file = "LICENSE"
 description = "A set of useful proc macros for wgpu"
 repository = "https://github.com/crowlKats/wgpu-macros"
 
-[lib]
-proc-macro = true
-
 [dependencies]
-quote = "1"
-proc-macro2 = "1.0"
-syn = { version = "1.0", features = ["default", "extra-traits"] }
-
-[dev-dependencies]
-wgpu = "0.12.0"
+wgpu-procs = { path = "wgpu-procs" }
+wgpu = "24.0.1"
