+++ wgpu-procs/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "wgpu-procs"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+quote = "1"
+proc-macro2 = "1.0"
+syn = { version = "1.0", features = ["default", "extra-traits"] }
